updated doc and sample codes with more comments.
