+++ cmd/selectmany/selectmany.go
@@ -1,3 +1,4 @@
+// This sample code demonstrates how to fetch many rows and allow cancel the query by Ctrl+C.
 package main
 
 import (
@@ -18,12 +19,12 @@ func main() {
 		flag.Parse()
 	}
 
+	// handler interrupt signal
 	ctx, cancel := context.WithCancel(context.Background())
 	c := make(chan os.Signal, 1)
 	signal.Notify(c, os.Interrupt)
 	defer func() {
 		signal.Stop(c)
-		cancel()
 	}()
 	go func() {
 		<-c
